📈Fixed memory issue
went from
592 in menu and
60 in game free memory to

682 in menu and
365 in game free memory

+++ matrix_boy/Breakout_game.cpp
@@ -1,6 +1,6 @@
 #include "ballgames.h"
 
-void Breakout::hit_log(int moveX, int moveY, int f = 0)
+void Breakout::hit_log(byte moveX, byte moveY, byte f = 0)
 {
     Serial.print("hits: ");
     Serial.print(hits);
@@ -32,7 +32,7 @@ void Breakout::hit_log(int moveX, int moveY, int f = 0)
     
 }
 
-bool Breakout::check_move(int nextX, int nextY)
+bool Breakout::check_move(byte nextX, byte nextY)
 {
     if(!(0 <= nextX && nextX < mapx))
     {
