+++ matrix_boy/ballgames.h
@@ -5,33 +5,33 @@
 
 class Ballgames {
     protected:
-    const int pad_size = 3;
+    const byte pad_size = 3;
 
-    const uint32_t ball_color = matrix.Color(255, 255, 255);
-    const uint32_t pad_color = matrix.Color(0, 0, 255);
+    const uint16_t ball_color = matrix.Color(255, 255, 255);
+    const uint16_t pad_color = matrix.Color(0, 0, 255);
     
 
-    const int bounce_vertical[4] = {3, 2, 1, 0};
-    const int bounce_horisontal[4] = {1, 0, 3, 2};
-    const int bounce_corner[4] = {2, 3, 0, 1};
+    const byte bounce_vertical[4] = {3, 2, 1, 0};
+    const byte bounce_horisontal[4] = {1, 0, 3, 2};
+    const byte bounce_corner[4] = {2, 3, 0, 1};
 
-    int pad_pos = 0;
+    byte pad_pos = 0;
 
     bool move_done;
 
-    int posx, posy, dir, speed;
+    byte posx, posy, dir, speed;
     bool run, gameover, win;
 
 };
 
 class Breakout: public Ballgames {
     private:
-    static const int rows = 8;
-    static const int coloums = mapx / 2;
+    static const byte rows = 8;
+    static const byte coloums = mapx / 2;
 
-    const int all_bricks = rows * coloums;
+    const byte all_bricks = rows * coloums;
 
-    const uint32_t row_colors[8] = {
+    const uint16_t row_colors[8] = {
         matrix.Color(255, 0, 0),
         matrix.Color(255, 0, 0),
         matrix.Color(255, 170, 0),
@@ -44,14 +44,14 @@ class Breakout: public Ballgames {
 
     bool bricks_map[coloums][rows];
     
-    bool check_move(int nextX, int nextY);
-    void hit_log(int nextX, int nextY, int f);
+    bool check_move(byte nextX, byte nextY);
+    void hit_log(byte nextX, byte nextY, byte f);
     void print();
 
     bool testmode = false;
     bool win = false;
 
-    int hits;
+    byte hits;
 
     public:
     void Breakout_game();
@@ -59,13 +59,13 @@ class Breakout: public Ballgames {
 
 class Pong: public Ballgames {
     private:
-    int enemy_pos;
-    const int enemy_size = 3;
-    const uint32_t enemy_color = matrix.Color(0, 0, 255);
-    int atack_cycle;
-    int required_atacks;
+    byte enemy_pos;
+    const byte enemy_size = 3;
+    const uint16_t enemy_color = matrix.Color(0, 0, 255);
+    uint16_t atack_cycle;
+    byte required_atacks;
 
-    bool check_pong_move(int nextX, int nextY);
+    bool check_pong_move(byte nextX, byte nextY);
 
     public:
     void Pong_game();
